+++ Analysis/plugins/ResolvedEMuCRanalyzer.cc
@@ -56,9 +56,9 @@ private:
   const edm::EDGetTokenT<edm::View<PileupSummaryInfo>> pileupToken_;
   const edm::EDGetTokenT<GenEventInfoProduct> generatorToken_;
   const edm::EDGetTokenT<reco::BeamSpot> beamspotToken_;
-  const edm::EDGetTokenT<double> prefweight_token;
-  const edm::EDGetTokenT<double> prefweightUp_token_;
-  const edm::EDGetTokenT<double> prefweightDn_token_;
+  const edm::EDGetTokenT<float> prefweight_token;
+  const edm::EDGetTokenT<float> prefweightUp_token_;
+  const edm::EDGetTokenT<float> prefweightDn_token_;
 
   const edm::EDGetTokenT<edm::TriggerResults> METfilterToken_;
   const edm::EDGetTokenT<edm::TriggerResults> triggerToken_;
@@ -115,9 +115,9 @@ genptcToken_(consumes<edm::View<reco::GenParticle>>(iConfig.getParameter<edm::In
 pileupToken_(consumes<edm::View<PileupSummaryInfo>>(iConfig.getParameter<edm::InputTag>("pileupSummary"))),
 generatorToken_(consumes<GenEventInfoProduct>(iConfig.getParameter<edm::InputTag>("generator"))),
 beamspotToken_(consumes<reco::BeamSpot>(iConfig.getParameter<edm::InputTag>("beamSpot"))),
-prefweight_token(consumes<double>(edm::InputTag("prefiringweight:nonPrefiringProb"))),
-prefweightUp_token_(consumes<double>(edm::InputTag("prefiringweight:nonPrefiringProbUp"))),
-prefweightDn_token_(consumes<double>(edm::InputTag("prefiringweight:nonPrefiringProbDown"))),
+prefweight_token(consumes<float>(edm::InputTag("prefiringweight:nonPrefiringProb"))),
+prefweightUp_token_(consumes<float>(edm::InputTag("prefiringweight:nonPrefiringProbUp"))),
+prefweightDn_token_(consumes<float>(edm::InputTag("prefiringweight:nonPrefiringProbDown"))),
 METfilterToken_(consumes<edm::TriggerResults>(iConfig.getParameter<edm::InputTag>("METfilters"))),
 triggerToken_(consumes<edm::TriggerResults>(iConfig.getParameter<edm::InputTag>("triggerResults"))),
 triggerobjectsToken_(consumes<edm::View<pat::TriggerObjectStandAlone>>(iConfig.getParameter<edm::InputTag>("triggerObjects"))),
@@ -387,20 +387,20 @@ void ResolvedEMuCRanalyzer::analyze(const edm::Event& iEvent, const edm::EventSe
   double purwgtUp = 1.;
   double purwgtDn = 1.;
   double purwgtNo = 1.;
-  double prefireNo = 1.;
-  double prefireUp = 1.;
-  double prefireDn = 1.;
+  float prefireNo = 1.;
+  float prefireUp = 1.;
+  float prefireDn = 1.;
 
   if (isMC_) {
-    edm::Handle<double> theprefweight;
+    edm::Handle<float> theprefweight;
     iEvent.getByToken(prefweight_token, theprefweight);
     prefireNo = *theprefweight;
 
-    edm::Handle<double> theprefweightUp;
+    edm::Handle<float> theprefweightUp;
     iEvent.getByToken(prefweightUp_token_, theprefweightUp);
     prefireUp = *theprefweightUp;
 
-    edm::Handle<double> theprefweightDn;
+    edm::Handle<float> theprefweightDn;
     iEvent.getByToken(prefweightDn_token_, theprefweightDn);
     prefireDn = *theprefweightDn;
